check both .my.cnf + mysql_password.txt

+++ nextcloud_update.sh
@@ -63,16 +63,19 @@ then
 fi
 
 # Make sure old instaces can upgrade as well
-if [ -f /var/mysql_password.txt ]
+if [ ! -f /root/.my.cnf ]
 then
-regressionpw=$(cat /var/mysql_password.txt)
-cat << LOGIN > "$MYCNF"
-[client]
-password='$regressionpw'
-LOGIN
-chmod 0600 $MYCNF
-chown root:root $MYCNF
-rm /var/mysql_password.txt
+    if [ -f /var/mysql_password.txt ]
+    then
+        regressionpw=$(cat /var/mysql_password.txt)
+        cat << LOGIN > "$MYCNF"
+        [client]
+        password='$regressionpw'
+        LOGIN
+        chmod 0600 $MYCNF
+        chown root:root $MYCNF
+        rm /var/mysql_password.txt
+    fi
 fi
 
 # Upgrade Nextcloud
